Be religious about using the cursor to access input
By accessing a reference to the slice from the enclosing scope, but
doing all modification throught he Cursor class, I confused myself as
far as what was the external and what was the internal representations,
and which would get passed where.

We had validation erros because we attempted to validate the input WITH
the cursor added, rather than without it. This commit moves some logic
around a little, adds a .Get method to the Cursor, and removes all
references to the input after the Cursor has been instantiated.